flatten fwd meta
