+++ stan/math/fwd/meta.hpp
@@ -1,9 +1,8 @@
 #ifndef STAN_MATH_FWD_META_HPP
 #define STAN_MATH_FWD_META_HPP
 
-#include <stan/math/fwd/scal/meta/is_fvar.hpp>
-#include <stan/math/fwd/scal/meta/partials_type.hpp>
-#include <stan/math/fwd/scal/meta/operands_and_partials.hpp>
-#include <stan/math/fwd/mat/meta/operands_and_partials.hpp>
+#include <stan/math/fwd/meta/is_fvar.hpp>
+#include <stan/math/fwd/meta/partials_type.hpp>
+#include <stan/math/fwd/meta/operands_and_partials.hpp>
 
 #endif
