+++ stan/math/fwd/meta/is_fvar.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_FWD_SCAL_META_IS_FVAR_HPP
-#define STAN_MATH_FWD_SCAL_META_IS_FVAR_HPP
+#ifndef STAN_MATH_FWD_META_IS_FVAR_HPP
+#define STAN_MATH_FWD_META_IS_FVAR_HPP
 
 #include <stan/math/fwd/core.hpp>
 #include <stan/math/prim/scal/meta/is_fvar.hpp>
