+++ stan/math/fwd/meta/partials_type.hpp
@@ -1,8 +1,8 @@
-#ifndef STAN_MATH_FWD_SCAL_META_PARTIALS_TYPE_HPP
-#define STAN_MATH_FWD_SCAL_META_PARTIALS_TYPE_HPP
+#ifndef STAN_MATH_FWD_META_PARTIALS_TYPE_HPP
+#define STAN_MATH_FWD_META_PARTIALS_TYPE_HPP
 
 #include <stan/math/fwd/core.hpp>
-#include <stan/math/fwd/scal/meta/is_fvar.hpp>
+#include <stan/math/fwd/meta/is_fvar.hpp>
 #include <stan/math/prim/scal/meta/partials_type.hpp>
 #include <type_traits>
