+++ stan/math/mix/meta.hpp
@@ -4,8 +4,8 @@
 #include <stan/math/prim/scal/meta/ad_promotable.hpp>
 
 #include <stan/math/fwd/core.hpp>
-#include <stan/math/fwd/scal/meta/is_fvar.hpp>
-#include <stan/math/fwd/scal/meta/partials_type.hpp>
+#include <stan/math/fwd/meta/is_fvar.hpp>
+#include <stan/math/fwd/meta/partials_type.hpp>
 
 #include <stan/math/rev/core.hpp>
 #include <stan/math/rev/scal/meta/is_var.hpp>
